exercise 4.21

+++ controllers/blogs.js
@@ -36,8 +36,23 @@ blogsRouter.post('/',  async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({error: 'invalid token'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  console.log(blog.user.toString(), decodedToken.id.toString())
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+
+  }
+
+  return response.status(400).json({error: 'user id does not match with blog id'})
+
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
